refactor: 즐겨찾기 단건 조회 결과 없을 시 404 리턴하도록 리팩터링
