+++ src/main/java/nextstep/common/ControllerExceptionHandler.java
@@ -1,5 +1,6 @@
-package nextstep.subway.ui;
+package nextstep.common;
 
+import nextstep.subway.exception.NotFoundFavoriteException;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
@@ -15,5 +16,10 @@ public class ControllerExceptionHandler {
     @ExceptionHandler(IllegalArgumentException.class)
     public ResponseEntity<Void> handleIllegalArgsException(IllegalArgumentException e) {
         return ResponseEntity.badRequest().build();
+    }
+
+    @ExceptionHandler(NotFoundFavoriteException.class)
+    public ResponseEntity<Void> handleIllegalArgsException(NotFoundFavoriteException e) {
+        return ResponseEntity.notFound().build();
     }
 }
